Move enums to separate files
