+++ web-api/src/cars/enums/body-type.enum.ts
@@ -1,0 +1,9 @@
+export enum BodyType {
+  SEDAN = "sedan",
+  SUV = "suv",
+  MINIVAN = "minivan",
+  PICKUP = "pickup",
+  HATCHBACK = "hatchback",
+  COUPE = "coupe",
+  STATION_WAGON = "station_wagon",
+}
