+++ web-api/src/cars/enums/drive-type.enum.ts
@@ -1,0 +1,5 @@
+export enum DriveType {
+  AWD = "awd",
+  FWD = "fwd",
+  RWD = "rwd",
+}
