+++ web-api/src/cars/enums/fuel-type.enum.ts
@@ -1,0 +1,4 @@
+export enum FuelType {
+  GASOLINE = "gasoline",
+  DIESEL = "diesel",
+}
