+++ web-api/src/cars/enums/transmission-type.enum.ts
@@ -1,0 +1,5 @@
+export enum TransmissionType {
+  AUTOMATIC = "automatic",
+  AUTOMATIC_CVT = "automatic_cvt",
+  MANUAL = "manual",
+}
